server: move getNodesBFS to utils/
